test(entity): keep migration runner regression isolated
